feat: rename icon
