+++ src/frontend/src/lib/components/icons/IconAstronautHelmet.svelte
